+++ src/main/java/survivalblock/atmosphere/atmospheric_api/not_mixin/entity/EntityWithAttributesImpl.java
@@ -9,7 +9,7 @@ import net.minecraft.nbt.NbtElement;
 
 import net.minecraft.storage.ReadView;
 import net.minecraft.storage.WriteView;
-///?}
+//?}
 import net.minecraft.world.World;
 
 import java.util.List;