+++ src/main/java/survivalblock/atmosphere/atmospheric_api/not_mixin/render/shader/client/AtmosphericShaderRegistryImpl.java
@@ -3,6 +3,7 @@ package survivalblock.atmosphere.atmospheric_api.not_mixin.render.shader.client;
 
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
+import net.minecraft.util.Identifier;
 import org.jetbrains.annotations.ApiStatus;
 
 import java.util.ArrayList;
@@ -20,6 +21,14 @@ public final class AtmosphericShaderRegistryImpl {
     static void register(ShaderApplier shaderApplier) {
         SHADER_APPLIERS.add(shaderApplier);
         SHADER_APPLIERS.sort(null);
+    }
+
+    static void register(Identifier id, ShaderApplier.PostEffectCondition condition) {
+        register(new ShaderApplier(id, condition));
+    }
+
+    static void register(Identifier texture, ShaderApplier.PostEffectCondition condition, int priority) {
+        register(new ShaderApplier(texture, condition, priority));
     }
 
     public static List<ShaderApplier> getShaderAppliers() {