+++ changelog.md
@@ -1,3 +1,3 @@
-Atmospheric API 3.0.3
-- Bump fabric API versions
-- Add DirectionalParticleS2CPayload
+Atmospheric API 3.1.0
+- Fix EntityWithAttributesImpl not existing
+- Fix pseudo-registry APIs